5_google_fit_agent/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
google-adk
pandas

